// spa/src/api/models/LaboratoryValueNameDropdownDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LaboratoryValueNameDropdownDTO
 */
export interface LaboratoryValueNameDropdownDTO {
    /**
     * 
     * @type {number}
     * @memberof LaboratoryValueNameDropdownDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LaboratoryValueNameDropdownDTO
     */
    name?: string;
}

/**
 * Check if a given object implements the LaboratoryValueNameDropdownDTO interface.
 */
export function instanceOfLaboratoryValueNameDropdownDTO(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function LaboratoryValueNameDropdownDTOFromJSON(json: any): LaboratoryValueNameDropdownDTO {
    return LaboratoryValueNameDropdownDTOFromJSONTyped(json, false);
}

export function LaboratoryValueNameDropdownDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): LaboratoryValueNameDropdownDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
    };
}

export function LaboratoryValueNameDropdownDTOToJSON(value?: LaboratoryValueNameDropdownDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
    };
}