// spa/src/api/models/LaboratoryValueRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LaboratoryValueRequest
 */
export interface LaboratoryValueRequest {
    /**
     * 
     * @type {number}
     * @memberof LaboratoryValueRequest
     */
    value?: number;
    /**
     * 
     * @type {number}
     * @memberof LaboratoryValueRequest
     */
    laboratoryValueNameId?: number;
}

/**
 * Check if a given object implements the LaboratoryValueRequest interface.
 */
export function instanceOfLaboratoryValueRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function LaboratoryValueRequestFromJSON(json: any): LaboratoryValueRequest {
    return LaboratoryValueRequestFromJSONTyped(json, false);
}

export function LaboratoryValueRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): LaboratoryValueRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'value': !exists(json, 'value') ? undefined : json['value'],
        'laboratoryValueNameId': !exists(json, 'laboratoryValueNameId') ? undefined : json['laboratoryValueNameId'],
    };
}

export function LaboratoryValueRequestToJSON(value?: LaboratoryValueRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'value': value.value,
        'laboratoryValueNameId': value.laboratoryValueNameId,
    };
}